+++ app/src/main/java/com/kasteca/activity/LogStudenteActivity.java
@@ -1,4 +1,4 @@
-package com.kasteca;
+package com.kasteca.activity;
 
         import androidx.annotation.NonNull;
         import androidx.appcompat.app.ActionBarDrawerToggle;
@@ -10,14 +10,15 @@ package com.kasteca;
 
         import android.content.Intent;
         import android.os.Bundle;
-        import android.view.Menu;
-        import android.view.MenuInflater;
         import android.view.MenuItem;
         import android.view.View;
         import android.widget.TextView;
 
         import com.google.android.material.navigation.NavigationView;
         import com.google.firebase.auth.FirebaseAuth;
+        import com.kasteca.fragment.CorsiDocenteFragment;
+        import com.kasteca.R;
+        import com.kasteca.object.Studente;
 
 public class LogStudenteActivity extends AppCompatActivity  implements NavigationView.OnNavigationItemSelectedListener {
