+++ app/src/main/java/com/kasteca/fragment/CorsiStudenteFragment.java
@@ -1,4 +1,4 @@
-package com.kasteca;
+package com.kasteca.fragment;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -6,6 +6,8 @@ import android.view.View;
 import android.view.ViewGroup;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+
+import com.kasteca.R;
 
 public class CorsiStudenteFragment extends Fragment {
