+++ app/src/main/java/com/kasteca/object/Commento.java
@@ -1,4 +1,4 @@
-package com.kasteca;
+package com.kasteca.object;
 
 import java.util.ArrayList;
 import java.util.Date;
