+++ app/src/main/java/com/kasteca/object/Risposta.java
@@ -1,4 +1,4 @@
-package com.kasteca;
+package com.kasteca.object;
 
 import java.util.Date;
