+++ app/src/main/res/layout/activity_log_docente.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".LogDocenteActivity"
+    tools:context=".activity.LogDocenteActivity"
     tools:openDrawer="start">
 
 <LinearLayout
