+++ app/src/main/res/layout/activity_log_studente.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".LogStudenteActivity"
+    tools:context=".activity.LogStudenteActivity"
     tools:openDrawer="start">
 
     <LinearLayout
