+++ app/src/main/res/layout/activity_registration.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorBackground"
-    tools:context=".RegistrationActivity">
+    tools:context=".activity.RegistrationActivity">
 
 
     <ImageView
