+++ app/src/main/res/layout/fragment_corsi_docente.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CorsiDocenteFragment" >
+    tools:context=".fragment.CorsiDocenteFragment" >
 
     <TextView
         android:id="@+id/text_home"
