+++ app/src/main/res/layout/fragment_corsi_studente.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CorsiStudenteFragment" >
+    tools:context=".fragment.CorsiStudenteFragment" >
 
     <TextView
         android:id="@+id/text_home"
